Make the test with a mocked server actually work
